Make spears don't get removed when used in creative.

+++ src/main/java/com/codetaylor/mc/pyrotech/modules/hunting/item/spi/ItemSpearBase.java
@@ -134,7 +134,7 @@ public abstract class ItemSpearBase
 
         world.playSound(null, player.posX, player.posY, player.posZ, SoundEvents.ENTITY_ARROW_SHOOT, SoundCategory.PLAYERS, 1.0F, 1.0F / (itemRand.nextFloat() * 0.4F + 1.2F) + velocity * 0.5F);
 
-        player.inventory.deleteStack(itemStack);
+        if (!player.isCreative()) player.inventory.deleteStack(itemStack);
 
         player.addStat(StatList.getObjectUseStats(this));
       }
